smart-access-control/backend/requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn==0.24.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0

# 监控
prometheus-client==0.19.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 代码质量
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 其他
redis==5.0.1
aioredis==2.0.1
